refactor(api): move listProposalLocations into its own router file
Per repo convention, each router lives in its own file merged via
mergeRouters. Extract the listProposalLocations procedure out of
listProposalsRouter into listProposalLocations.ts and register it in the
proposals router index.
