+++ services/api/src/routers/decision/proposals/index.ts
@@ -8,6 +8,7 @@ import { getExportStatusRouter } from './getExportStatus';
 import { getLatestSelectionForProposalRouter } from './getLatestSelection';
 import { getProposalWithReviewAggregatesRouter } from './getProposalWithReviewAggregates';
 import { listProposalsRouter } from './list';
+import { listProposalLocationsRouter } from './listProposalLocations';
 import { listWithReviewAggregatesRouter } from './listWithReviewAggregates';
 import { submitProposalRouter } from './submit';
 import { updateProposalRouter } from './update';
@@ -19,6 +20,7 @@ export const proposalsRouter = mergeRouters(
   getLatestSelectionForProposalRouter,
   getProposalWithReviewAggregatesRouter,
   listProposalsRouter,
+  listProposalLocationsRouter,
   listWithReviewAggregatesRouter,
   submitProposalRouter,
   updateProposalRouter,
