+++ services/api/src/routers/decision/proposals/list.ts
@@ -1,20 +1,11 @@
-import {
-  Channels,
-  listAllProposals,
-  listProposalLocations,
-  listProposals,
-} from '@op/common';
+import { Channels, listAllProposals, listProposals } from '@op/common';
 import {
   allProposalsFilterSchema,
   allProposalsListSchema,
   proposalListSchema,
-  proposalLocationsSchema,
 } from '@op/common/client';
 
-import {
-  proposalFilterSchema,
-  proposalLocationsFilterSchema,
-} from '../../../encoders/decision';
+import { proposalFilterSchema } from '../../../encoders/decision';
 import { openProcedure, router } from '../../../trpcFactory';
 
 export const listProposalsRouter = router({
@@ -51,25 +42,5 @@ export const listProposalsRouter = router({
       ]);
 
       return allProposalsListSchema.parse(result);
-    }),
-  /**
-   * Every located proposal in the instance's current scope — the map's pin
-   * source, so the map isn't capped by the list's page size.
-   */
-  listProposalLocations: openProcedure()
-    .input(proposalLocationsFilterSchema)
-    .output(proposalLocationsSchema)
-    .query(async ({ ctx, input }) => {
-      const { user } = ctx;
-      const result = await listProposalLocations({
-        input,
-        user,
-      });
-
-      ctx.registerQueryChannels([
-        Channels.decisionProposals(input.processInstanceId),
-      ]);
-
-      return proposalLocationsSchema.parse(result);
     }),
 });
